requirements.txt
# Core deep learning
torch>=1.9.0
torchvision>=0.10.0
pytorch-lightning>=1.5.0

# Computer vision and segmentation
segmentation-models-pytorch>=0.3.0
transformers>=4.0.0
opencv-python>=4.5.0
pillow>=8.0.0
scikit-image>=0.18.0

# Scientific computing
numpy>=1.20.0
scipy>=1.7.0

# Geospatial and data handling
rasterio>=1.2.0
netCDF4>=1.5.0
gdal>=3.0.0

# Medical image processing
medpy>=0.4.0

# Visualization and plotting
matplotlib>=3.3.0

# Utilities
tqdm>=4.60.0
pyyaml>=5.4.0
pathlib>=1.0.1

# Experiment tracking
wandb>=0.12.0

# Development tools (optional)
jupyter>=1.0.0
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910